requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
click==7.1.2
cycler==0.10.0
dataclasses==0.6
Flask==1.1.2
future==0.18.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.17.0
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.3
numpy==1.19.4
olefile==0.46
opencv-python==4.4.0.46
pandas==1.1.4
Pillow==8.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
scikit-learn==0.23.2
scipy==1.5.4
threadpoolctl==2.1.0
torch==1.7.0+cpu
torchvision==0.8.1+cpu
Werkzeug==1.0.1
